@@ -18,17 +18,16 @@
   },
   "private": false,
   "dependencies": {
-    "@angular/animations": "^7.2.9",
-    "@angular/cdk": "^7.2.1",
-    "@angular/common": "~7.2.9",
-    "@angular/compiler": "~7.2.9",
-    "@angular/core": "~7.2.9",
-    "@angular/forms": "~7.2.9",
-    "@angular/http": "~7.2.9",
-    "@angular/material": "^7.2.0",
-    "@angular/platform-browser": "~7.2.9",
-    "@angular/platform-browser-dynamic": "~7.2.9",
-    "@angular/router": "~7.2.9",
+    "@angular/animations": "^8.0.2",
+    "@angular/cdk": "^8.0.1",
+    "@angular/common": "~8.0.2",
+    "@angular/compiler": "~8.0.2",
+    "@angular/core": "~8.0.2",
+    "@angular/forms": "~8.0.2",
+    "@angular/material": "^8.0.1",
+    "@angular/platform-browser": "~8.0.2",
+    "@angular/platform-browser-dynamic": "~8.0.2",
+    "@angular/router": "~8.0.2",
     "@types/quill": "^1.3.6",
     "ag-grid-angular": "^20.0.0",
     "ag-grid-community": "^20.0.0",
@@ -37,49 +36,50 @@
     "date-fns": "^1.30.1",
     "hammerjs": "^2.0.8",
     "lodash": "^4.17.11",
-    "ng-mocks": "^7.5.0",
+    "ng-mocks": "^8.0.0",
     "quill": "^1.3.6",
-    "rxjs": "~6.4.0",
-    "zone.js": "~0.8.29"
+    "rxjs": "~6.5.2",
+    "zone.js": "~0.9.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.12.1",
-    "@angular-devkit/build-ng-packagr": "~0.12.1",
-    "@angular-devkit/core": "^7.3.6",
-    "@angular/cli": "~7.2.1",
-    "@angular/compiler-cli": "~7.2.9",
-    "@angular/language-service": "~7.2.9",
+    "@angular-devkit/build-angular": "~0.800.3",
+    "@angular-devkit/build-ng-packagr": "~0.800.3",
+    "@angular-devkit/core": "^8.0.3",
+    "@angular/cli": "~8.0.3",
+    "@angular/compiler-cli": "~8.0.2",
+    "@angular/language-service": "~8.0.2",
     "@babel/core": "^7.2.2",
-    "@storybook/addon-actions": "^5.0.10",
-    "@storybook/addon-knobs": "^5.0.10",
-    "@storybook/addon-links": "^5.0.10",
-    "@storybook/addon-notes": "^5.0.10",
-    "@storybook/addons": "^5.0.10",
-    "@storybook/angular": "^5.0.10",
+    "@storybook/addon-actions": "^5.1.9",
+    "@storybook/addon-knobs": "^5.1.9",
+    "@storybook/addon-links": "^5.1.9",
+    "@storybook/addon-notes": "^5.1.9",
+    "@storybook/addons": "^5.1.9",
+    "@storybook/angular": "^5.1.9",
     "@storybook/storybook-deployer": "^2.8.1",
     "@types/jasmine": "~2.8.8",
     "@types/jasminewd2": "~2.0.3",
     "@types/node": "^8.9.5",
     "babel-loader": "^8.0.5",
-    "codelyzer": "~4.5.0",
+    "codelyzer": "^5.0.1",
     "cpx": "^1.5.0",
     "jasmine-core": "~2.99.1",
-    "jasmine-marbles": "^0.4.1",
+    "jasmine-marbles": "^0.6.0",
     "jasmine-spec-reporter": "~4.2.1",
     "karma": "~3.0.0",
     "karma-chrome-launcher": "~2.2.0",
     "karma-coverage-istanbul-reporter": "~2.0.1",
     "karma-jasmine": "~1.1.2",
     "karma-jasmine-html-reporter": "^0.2.2",
-    "ng-packagr": "^4.7.1",
+    "ng-packagr": "^5.1.0",
+    "node-sass": "^4.12.0",
+    "prettier": "^1.16.4",
     "protractor": "~5.4.1",
     "ts-node": "~7.0.1",
     "tsconfig-paths": "^3.7.0",
-    "tsickle": ">=0.29.0",
-    "tslib": "^1.9.0",
+    "tsickle": "^0.35.0",
+    "tslib": "^1.10.0",
     "tslint": "~5.12.1",
-    "typescript": "~3.2.4",
-    "prettier": "^1.16.4"
+    "typescript": "~3.4.5"
   },
   "publishConfig": {
     "registry": "https://hibob.jfrog.io/hibob/api/npm/bob-js-virtual/"
