@@ -1,12 +1,12 @@
 {
   "name": "bob-style",
-  "version": "1.1.100",
+  "version": "1.1.101",
   "peerDependencies": {
-    "@angular/animations": "^7.2.0",
-    "@angular/cdk": "^7.2.1",
-    "@angular/common": "^7.2.0",
-    "@angular/core": "^7.2.0",
-    "@angular/material": "^7.2.1",
+    "@angular/animations": "^8.0.2",
+    "@angular/cdk": "^8.0.1",
+    "@angular/common": "^8.0.2",
+    "@angular/core": "^8.0.2",
+    "@angular/material": "^8.0.1",
     "lodash": "^4.17.11",
     "hammerjs": "^2.0.8",
     "@types/node": "^8.9.5",
