@@ -3,7 +3,7 @@ import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 import { ButtonComponent } from './button.component';
 import { ButtonSize, ButtonType } from '../buttons.enum';
 import { By } from '@angular/platform-browser';
-import { MatButtonModule } from '@angular/material';
+import { MatButtonModule } from '@angular/material/button';
 import { IconsModule } from '../../../icons/icons.module';
 import { IconService } from '../../../icons/icon.service';
 import { IconColor, Icons, IconSize } from '../../../icons/icons.enum';
