@@ -1,6 +1,6 @@
 import {NgModule} from '@angular/core';
 import {SliderComponent} from './slider.component';
-import {MatSliderModule} from '@angular/material';
+import { MatSliderModule } from '@angular/material/slider';
 import {CommonModule} from '@angular/common';
 import 'hammerjs';
 
