@@ -1,20 +1,12 @@
 import { storiesOf } from '@storybook/angular';
-import {
-  text,
-  select,
-  boolean,
-  withKnobs
-} from '@storybook/addon-knobs/angular';
+import { text, select, boolean, withKnobs } from '@storybook/addon-knobs/angular';
 import { action } from '@storybook/addon-actions';
 import { SwitchToggleModule } from './switch-toggle.module';
 import { values } from 'lodash';
 import { ComponentGroupType } from '../../consts';
 import { StoryBookLayoutModule } from '../../story-book-layout/story-book-layout.module';
 
-const buttonStories = storiesOf(
-  ComponentGroupType.ButtonsAndIndicators,
-  module
-).addDecorator(withKnobs);
+const buttonStories = storiesOf(ComponentGroupType.ButtonsAndIndicators, module).addDecorator(withKnobs);
 
 const template = `
 <b-switch-toggle [isDisabled]="isDisabled"
@@ -55,7 +47,7 @@ buttonStories.add(
     props: {
       isDisabled: boolean('isDisabled', false),
       isChecked: boolean('isChecked', true),
-      switchChange: action()
+      switchChange: action('switchChange')
     },
     moduleMetadata: {
       imports: [SwitchToggleModule, StoryBookLayoutModule]
