@@ -1,11 +1,5 @@
 import { storiesOf } from '@storybook/angular';
-import {
-  text,
-  select,
-  boolean,
-  withKnobs,
-  number
-} from '@storybook/addon-knobs/angular';
+import { text, select, boolean, withKnobs, number } from '@storybook/addon-knobs/angular';
 import { action } from '@storybook/addon-actions';
 import { values } from 'lodash';
 import { TextareaModule } from './textarea.module';
@@ -13,10 +7,7 @@ import { ComponentGroupType } from '../../consts';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { StoryBookLayoutModule } from '../../story-book-layout/story-book-layout.module';
 
-const textareaStories = storiesOf(
-  ComponentGroupType.FormElements,
-  module
-).addDecorator(withKnobs);
+const textareaStories = storiesOf(ComponentGroupType.FormElements, module).addDecorator(withKnobs);
 
 const template = `
 <b-textarea [maxChars]="maxChars"
@@ -67,7 +58,7 @@ textareaStories.add(
     return {
       template: storyTemplate,
       props: {
-        inputEvents: action(),
+        inputEvents: action('inputEvents'),
         maxChars: number('maxChars', ''),
         value: text('value', ''),
         label: text('label', 'Input label'),
@@ -79,11 +70,7 @@ textareaStories.add(
         errorMessage: text('errorMessage', '')
       },
       moduleMetadata: {
-        imports: [
-          BrowserAnimationsModule,
-          TextareaModule,
-          StoryBookLayoutModule
-        ]
+        imports: [BrowserAnimationsModule, TextareaModule, StoryBookLayoutModule]
       }
     };
   },
