@@ -1,5 +1,5 @@
 import { Injectable } from '@angular/core';
-import { MatIconRegistry } from '@angular/material';
+import { MatIconRegistry } from '@angular/material/icon';
 import { DomSanitizer } from '@angular/platform-browser';
 import { Icons } from './icons.enum';
 import { first } from 'rxjs/operators';
