@@ -6,7 +6,7 @@ import { TypographyModule } from '../../typography/typography.module';
 import { UtilsService } from '../../services/utils/utils.service';
 import { UtilsModule } from '../../services/utils/utils.module';
 import { ButtonsModule } from '../../buttons-indicators/buttons/buttons.module';
-import { MatTooltipModule } from '@angular/material';
+import { MatTooltipModule } from '@angular/material/tooltip';
 
 @NgModule({
   declarations: [BreadcrumbsComponent],
