@@ -1,13 +1,5 @@
 import { storiesOf } from '@storybook/angular';
-import {
-  array,
-  boolean,
-  number,
-  object,
-  select,
-  text,
-  withKnobs
-} from '@storybook/addon-knobs/angular';
+import { array, boolean, number, object, select, text, withKnobs } from '@storybook/addon-knobs/angular';
 import { action } from '@storybook/addon-actions';
 import { ComponentGroupType } from '../../consts';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
@@ -15,10 +7,7 @@ import { BreadcrumbsModule } from './breadcrumbs.module';
 import { Breadcrumb, BreadcrumbNavButtons } from './breadcrumbs.interface';
 import { StoryBookLayoutModule } from '../../story-book-layout/story-book-layout.module';
 
-const inputStories = storiesOf(
-  ComponentGroupType.Navigation,
-  module
-).addDecorator(withKnobs);
+const inputStories = storiesOf(ComponentGroupType.Navigation, module).addDecorator(withKnobs);
 
 const componmentTemplate = `
 <b-breadcrumbs [breadcrumbs]="breadcrumbs"
@@ -80,16 +69,12 @@ inputStories.add(
         breadcrumbs: object<Breadcrumb>('breadcrumbs', breadcrumbsMock),
         buttons: object<BreadcrumbNavButtons>('buttons', breadcrumbsButtons),
         activeIndex: number('activeIndex', 2),
-        stepClick: action(),
-        nextClick: action(),
-        prevClick: action()
+        stepClick: action('stepClick'),
+        nextClick: action('nextClick'),
+        prevClick: action('prevClick')
       },
       moduleMetadata: {
-        imports: [
-          BrowserAnimationsModule,
-          BreadcrumbsModule,
-          StoryBookLayoutModule
-        ]
+        imports: [BrowserAnimationsModule, BreadcrumbsModule, StoryBookLayoutModule]
       }
     };
   },
