@@ -8,33 +8,24 @@ export function MixIn(baseCtors: Function[]) {
   };
 }
 
-export const randomFromArray = (array: any[], num: number = 1) =>
-  array.sort(() => 0.5 - Math.random()).slice(0, num);
+export const randomFromArray = (array: any[], num: number = 1) => array.sort(() => 0.5 - Math.random()).slice(0, num);
 
 export const keysFromArrayOrObject = (smth: string[] | {}): string[] =>
   Array.isArray(smth) ? smth : Object.keys(smth);
 
-export const getKeyByValue = (object: object, value: any) =>
-  Object.keys(object).find(key => object[key] === value);
+export const getKeyByValue = (object: object, value: any) => Object.keys(object).find((key) => object[key] === value);
 
 export const isString = (val: any): boolean => val && typeof val === 'string';
 
 export const isArray = (val: any): boolean => val && Array.isArray(val);
 
-export const isObject = (val: any): boolean =>
-  val && !isArray(val) && val === Object(val);
+export const isObject = (val: any): boolean => val && !isArray(val) && val === Object(val);
 
-export const isNullOrUndefined = (val: any): boolean =>
-  val === undefined || val === null;
+export const isNullOrUndefined = (val: any): boolean => val === undefined || val === null;
 
-export const isRenderedComponent = (obj: any): boolean =>
-  obj && !!obj.component;
+export const isRenderedComponent = (obj: any): boolean => obj && !!obj.component;
 
-export const simpleUID = (
-  prefix: string = '',
-  length: number = 5,
-  suffix: string = ''
-): string => {
+export const simpleUID = (prefix: string = '', length: number = 5, suffix: string = ''): string => {
   return (
     prefix +
     Math.random()
@@ -46,8 +37,7 @@ export const simpleUID = (
 
 export const pass = (a: any): any => a;
 
-export const isKey = (key: string, expected: string): boolean =>
-  key.toUpperCase() === expected.toUpperCase();
+export const isKey = (key: string, expected: string): boolean => key.toUpperCase() === expected.toUpperCase();
 
 export const stringListToArray = (list: string): string[] => {
   if (isArray(list) || !list) {
@@ -56,23 +46,16 @@ export const stringListToArray = (list: string): string[] => {
   if (!isString(list)) {
     return [list];
   }
-  return Array.from(
-    new Set(list.split(/[^\w\u0020]+/).map(i => i.trim()))
-  ).filter(i => !!i);
+  return Array.from(new Set(list.split(/[^\w\u0020]+/).map((i) => i.trim()))).filter((i) => !!i);
 };
 
-export const asArray = (smth: any): any[] =>
-  !isNullOrUndefined(smth) ? (isArray(smth) ? smth : [smth]) : [];
+export const asArray = (smth: any): any[] => (!isNullOrUndefined(smth) ? (isArray(smth) ? smth : [smth]) : []);
 
 export const asNumber = (smth: any): number => parseInt(smth, 10);
 
-export const compareAsNumbers = (
-  a: string | number,
-  b: string | number
-): boolean => asNumber(a) === asNumber(b);
+export const compareAsNumbers = (a: string | number, b: string | number): boolean => asNumber(a) === asNumber(b);
 
-export const compareAsStrings = (a: any, b: any): boolean =>
-  String(a) === String(b);
+export const compareAsStrings = (a: any, b: any): boolean => String(a) === String(b);
 
 export const countChildren = (parentSelector, parent) => {
   parent = parentSelector ? document.querySelector(parentSelector) : parent;
@@ -92,47 +75,28 @@ export const countChildren = (parentSelector, parent) => {
 };
 
 import {
-  isPlainObject as _isPlainObject,
+  compose as _compose,
   isArray as _isArray,
-  reduce as _reduce,
+  isPlainObject as _isPlainObject,
   merge as _merge,
-  toPairs as _toPairs,
+  reduce as _reduce,
   set as _set,
-  compose as _compose
+  toPairs as _toPairs
 } from 'lodash/fp';
 
 export const flatten = (obj, path = []) => {
   return _isPlainObject(obj) || _isArray(obj)
-    ? _reduce(
-        (acc, [k, v]) => _merge(acc, flatten(v, [...path, k])),
-        {},
-        _toPairs(obj)
-      )
+    ? _reduce((acc, [k, v]) => _merge(acc, flatten(v, [...path, k])), {}, _toPairs(obj))
     : { [path.join('.')]: obj };
 };
 
-export const unflatten = _compose(
-  _reduce((acc, [k, v]) => _set(k, v, acc), {}),
-  _toPairs
-);
+export const unflatten = _compose(_reduce((acc, [k, v]) => _set(k, v, acc), {}), _toPairs);
 
 export const stringify = (smth: any): string =>
-  isString(smth)
-    ? smth
-    : isArray(smth)
-    ? smth.join(', ')
-    : isObject(smth)
-    ? JSON.stringify(smth)
-    : String(smth);
+  isString(smth) ? smth : isArray(smth) ? smth.join(', ') : isObject(smth) ? JSON.stringify(smth) : String(smth);
 
-export const getType = smth =>
-  smth === null
-    ? 'null'
-    : isArray(smth)
-    ? 'array'
-    : smth instanceof Date
-    ? 'date'
-    : String(typeof smth);
+export const getType = (smth) =>
+  smth === null ? 'null' : isArray(smth) ? 'array' : smth instanceof Date ? 'date' : String(typeof smth);
 
 export const domainFromUrl = (url: string) => {
   let a = document.createElement('a');
